Fix checkstyle and move the private confs to SnapshotManager.
